Extract step 1 specific styles to separate stylesheet
