+++ client/dashboard/profile-wizard/steps/start/images/flash_on.js
