+++ client/dashboard/profile-wizard/steps/start/images/local_atm.js
