+++ client/dashboard/profile-wizard/steps/start/images/phone_android.js
