+++ client/dashboard/profile-wizard/steps/start/images/security.js
